Create Main
Enter a number of rows and columns, and it will print out how it should look like. Make sure you've also downloaded and added the Place, Light, and Road classes.

You might need to change the view settings for the console in Eclipse or else it might look weird. Right click the console and make sure word-wrap is off. And right click the console, preferences, and uncheck the 'fixed console width' box.