Update Map.java
carSpawn() now works properly! I had to fix my maths a bit...

deleteCars() has replaced deleteCar(Car c) to delete all cars outside the grid: the issue before was that when removing cars from the arraylist before, it would move each subsequent element one space to the left in the array, which would skip some through the rest of the pass: now it resets the iterator to 0 to rescan the array when a deletion takes place.
deleteCars() is only called when move() tells the main method a car is outside the array.